move create elements functions from aframe-streetmix-parser.js to create-elements.js file
